Fix application_status telemetry data
It turns out there is only one application_status value,
and the in position bit means all actuators are in position.
I thought application status was per-actuator, since there are 6 values
but in fact the first of these values is the true application status
and the other 5 values are always zero.
Those extra values will go away in DM-28998 (for the low-level controller)
and a corresponding update for this package.