HexapodCsc: bug fix: move and offset commands
Allow move and offset commands while moving. There was already
code to stop the current move but it was never run because
of an incorrect initial assert.

Also update the unit test of interruption to make sure one move
has started before starting the next. The moves were happening
so quickly that the above bug was missed.