Update the Telemetry struct
And update the code to use the modified struct.
Most of the required updates are because position value are now broken
into two sets of three: xyz and uvw, because the EUI order is uvwxyz.
Note: I moved motor current and voltage to the end of the Telemetry
struct because that is where Te-Wei will likely choose to put it
(he will almost certainly not mix it in with values that the EUI uses,
to avoid having to compute new offsets for the various fields).