In hexapod_csc.py, update compensation_loop to send the CSC to Fault if compensation fails.
This current behavior is actually not correct, CSCs should go to Fault is something unnexpected happens. This was causing issues at night as we were operating for long periods with compensation off and having it get enabled by a slew operation, causing the hexapod to jump position. See OBS-648 for more information.